+++ crates/sic_image_engine/src/engine.rs
@@ -122,8 +122,8 @@ impl ImageEngine {
             }
             ImgOp::Diff(path) => operations::diff::Diff::new(path).apply_operation(&mut self.image),
             ImgOp::Dither => operations::dither::Dither::new().apply_operation(&mut self.image),
-            ImgOp::DitherColor((colors, sample_factor)) => {
-                operations::dither_color::DitherColor::new(*colors, *sample_factor)
+            ImgOp::DitherQuant((colors, sample_factor)) => {
+                operations::dither_quant::DitherQuant::new(*colors, *sample_factor)
                     .apply_operation(&mut self.image)
             }
             ImgOp::DrawText(inner) => {
@@ -143,6 +143,13 @@ impl ImageEngine {
             }
             ImgOp::HueRotate(degree) => {
                 operations::hue_rotate::HueRotate::new(*degree).apply_operation(&mut self.image)
+            }
+            ImgOp::IndexColors => {
+                operations::index_colors::IndexColors::new().apply_operation(&mut self.image)
+            }
+            ImgOp::IndexColorsQuant((colors, sample_factor)) => {
+                operations::index_colors_quant::IndexColorsQuant::new(*colors, *sample_factor)
+                    .apply_operation(&mut self.image)
             }
             ImgOp::HorizontalGradient(colors) => {
                 operations::horizontal_gradient::HorizontalGradient::new(*colors)
@@ -1320,11 +1327,11 @@ mod tests {
     }
 
     #[test]
-    fn test_dither_color() {
+    fn test_dither_quant() {
         let img = setup_default_test_image();
         let cmp = setup_default_test_image();
 
-        let operation = ImgOp::DitherColor((64, 10));
+        let operation = ImgOp::DitherQuant((64, 10));
 
         let operator = ImageEngine::new(img);
         let done = operator.ignite(&[Instr::Operation(operation)]);
@@ -1334,34 +1341,100 @@ mod tests {
 
         assert_ne!(cmp.raw_pixels(), result_img.raw_pixels());
 
-        output_test_image_for_manual_inspection(&result_img, out_!("test_dither_color.png"));
+        output_test_image_for_manual_inspection(&result_img, out_!("test_dither_quant.png"));
     }
 
     #[test]
-    fn test_dither_color_colors_out_of_range() {
+    fn test_dither_quant_colors_out_of_range() {
         let img = setup_default_test_image();
 
-        let operation = ImgOp::DitherColor((32, 10));
+        let operation = ImgOp::DitherQuant((32, 10));
 
         let operator = ImageEngine::new(img);
         let done = operator.ignite(&[Instr::Operation(operation)]);
         assert!(matches!(
             done,
-            Err(SicImageEngineError::DitherColorsOutOfRange(32))
+            Err(SicImageEngineError::DitherQuantColorsOutOfRange(32))
         ));
     }
 
     #[test]
-    fn test_dither_color_sample_factor_out_of_range() {
+    fn test_dither_quant_sample_factor_out_of_range() {
         let img = setup_default_test_image();
 
-        let operation = ImgOp::DitherColor((64, 0));
+        let operation = ImgOp::DitherQuant((64, 0));
 
         let operator = ImageEngine::new(img);
         let done = operator.ignite(&[Instr::Operation(operation)]);
         assert!(matches!(
             done,
-            Err(SicImageEngineError::DitherSampleFactorOutOfRange(0))
+            Err(SicImageEngineError::DitherQuantSampleFactorOutOfRange(0))
+        ));
+    }
+
+    #[test]
+    fn test_index_colors() {
+        let img = setup_default_test_image();
+        let cmp = setup_default_test_image();
+
+        let operation = ImgOp::IndexColors;
+
+        let operator = ImageEngine::new(img);
+        let done = operator.ignite(&[Instr::Operation(operation)]);
+        assert!(done.is_ok());
+
+        let result_img = done.unwrap();
+
+        assert_ne!(cmp.raw_pixels(), result_img.raw_pixels());
+
+        output_test_image_for_manual_inspection(&result_img, out_!("test_index_colors.png"));
+    }
+
+    #[test]
+    fn test_index_colors_quant() {
+        let img = setup_default_test_image();
+        let cmp = setup_default_test_image();
+
+        let operation = ImgOp::IndexColorsQuant((64, 10));
+
+        let operator = ImageEngine::new(img);
+        let done = operator.ignite(&[Instr::Operation(operation)]);
+        assert!(done.is_ok());
+
+        let result_img = done.unwrap();
+
+        assert_ne!(cmp.raw_pixels(), result_img.raw_pixels());
+
+        output_test_image_for_manual_inspection(&result_img, out_!("test_index_colors_quant.png"));
+    }
+
+    #[test]
+    fn test_index_colors_quant_colors_out_of_range() {
+        let img = setup_default_test_image();
+
+        let operation = ImgOp::IndexColorsQuant((32, 10));
+
+        let operator = ImageEngine::new(img);
+        let done = operator.ignite(&[Instr::Operation(operation)]);
+        assert!(matches!(
+            done,
+            Err(SicImageEngineError::IndexColorsQuantColorsOutOfRange(32))
+        ));
+    }
+
+    #[test]
+    fn test_index_colors_quant_sample_factor_out_of_range() {
+        let img = setup_default_test_image();
+
+        let operation = ImgOp::IndexColorsQuant((64, 0));
+
+        let operator = ImageEngine::new(img);
+        let done = operator.ignite(&[Instr::Operation(operation)]);
+        assert!(matches!(
+            done,
+            Err(SicImageEngineError::IndexColorsQuantSampleFactorOutOfRange(
+                0
+            ))
         ));
     }