+++ crates/sic_image_engine/src/errors.rs
@@ -29,10 +29,18 @@ pub enum SicImageEngineError {
     BlurSigmaNotPositive(f32),
 
     #[error("Unable to dither; the number of colors must be >= 64, but was '{0}'")]
-    DitherColorsOutOfRange(u32),
+    DitherQuantColorsOutOfRange(u32),
 
     #[error("Unable to dither; the sample factor must be in the range [1, 30], but was '{0}'")]
-    DitherSampleFactorOutOfRange(u32),
+    DitherQuantSampleFactorOutOfRange(u32),
+
+    #[error("Unable to index colors; the number of colors must be >= 64, but was '{0}'")]
+    IndexColorsQuantColorsOutOfRange(u32),
+
+    #[error(
+        "Unable to index colors; the sample factor must be in the range [1, 30], but was '{0}'"
+    )]
+    IndexColorsQuantSampleFactorOutOfRange(u32),
 
     #[error("Unable to load font: '{0}'")]
     FontError(sic_core::ab_glyph::InvalidFont),